requirements.txt
mlflow==2.10.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
flask==2.3.2
gunicorn==20.1.0
prometheus-client==0.17.0
requests==2.32.0
pytest==7.4.0
python-dotenv==1.0.0
joblib==1.3.2
evidently==0.5.0
pydantic==1.10.15
slack-sdk==3.21.3
boto3==1.34.0
minio==7.1.15
gitpython==3.1.40
psutil==5.9.0